move functions into modules, fix imports
