+++ parlai/agents/transformer/modules/__init__.py
@@ -3,9 +3,15 @@
 # This source code is licensed under the MIT license found in the
 # LICENSE file in the root directory of this source tree.
 
-from parlai.agents.transformer.modules.generator import *  # noqa: F401
-from parlai.agents.transformer.modules.mem_net import *  # noqa: F401
-from parlai.agents.transformer.modules.decoder import *  # noqa: F401
-from parlai.agents.transformer.modules.encoder import *  # noqa: F401
-from parlai.agents.transformer.modules.attention import *  # noqa: F401
-from parlai.agents.transformer.modules.layer_norm import *  # noqa: F401
+from .functions import create_embeddings, create_position_codes  # noqa: F401
+from .attention import BasicAttention, MultiHeadAttention  # noqa: F401
+from .ffn import TransformerFFN  # noqa: F401
+from .encoder import TransformerEncoder, TransformerEncoderLayer  # noqa: F401
+from .decoder import TransformerDecoder, TransformerDecoderLayer  # noqa: F401
+from .generator import TransformerGeneratorModel  # noqa: F401
+from .mem_net import (  # noqa: F401
+    get_n_positions_from_options,
+    TransformerLinearWrapper,
+    TransformerMemNetModel,
+    TransformerResponseWrapper,
+)
