+++ parlai/agents/transformer/modules/functions.py
