+++ parlai/agents/transformer/modules/mem_net.py
@@ -5,8 +5,11 @@
 
 import torch.nn as nn
 
-from parlai.agents.transformer.functions import create_embeddings
-from parlai.agents.transformer.modules import BasicAttention, TransformerEncoder
+from parlai.agents.transformer.modules import (
+    create_embeddings,
+    BasicAttention,
+    TransformerEncoder,
+)
 
 
 def get_n_positions_from_options(opt):
